+++ protocol/src/lib.rs
@@ -1,10 +1,10 @@
 #[derive(Debug, Copy, Clone)]
 pub struct Position {
-    pub x: f32,
-    pub y: f32,
+    pub x: i32,
+    pub y: i32,
 }
 impl Position {
-    pub fn new(x: f32, y: f32) -> Self {
+    pub fn new(x: i32, y: i32) -> Self {
         Self { x, y }
     }
 }
